requirements.txt
flask>=3.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
requests>=2.31.0
Pillow>=10.0.0
tensorflow>=2.15.0
bcrypt>=4.0.0
bleach>=6.0.0
